quete 17.2

+++ src/Controller/ArticleController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Article;
+use App\Entity\User;
 use App\Form\Article1Type;
 use App\Service\Slugify;
 use App\Repository\ArticleRepository;
@@ -10,6 +11,8 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+
 
 
 /**
@@ -22,8 +25,10 @@ class ArticleController extends AbstractController
      */
     public function index(ArticleRepository $articleRepository): Response
     {   
+
         return $this->render('article/index.html.twig', [
             'articles' => $articleRepository->findAll(),
+
         ]);
     }
 
@@ -39,6 +44,8 @@ class ArticleController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
             $article->setSlug($slugify->generate($article->getTitle()));
+            $author = $this->getUser();
+            $article->setAuthor($author);
             $entityManager->persist($article);
             $entityManager->flush();
 
@@ -53,19 +60,23 @@ class ArticleController extends AbstractController
 
     /**
      * @Route("/{id}", name="article_show", methods={"GET"})
+     * @IsGranted("ROLE_AUTHOR")
      */
     public function show(Article $article): Response
-    {   
+    {  
         return $this->render('article/show.html.twig', [
             'article' => $article,
         ]);
     }
 
     /**
-     * @Route("/{id}/edit", name="article_edit", methods={"GET","POST"})
+     * @Route("/{id}/edit", name="article_edit", methods={"GET","POST"})    
+     * @IsGranted("ROLE_AUTHOR")
      */
-    public function edit(Request $request, Article $article): Response
-    {
+    public function edit(Request $request, Article $article, Slugify $slugify): Response
+    {  
+        $this->denyAccessUnlessGranted('EDIT', $article);
+
         $form = $this->createForm(Article1Type::class, $article);
         $form->handleRequest($request);
 
@@ -86,9 +97,12 @@ class ArticleController extends AbstractController
 
     /**
      * @Route("/{id}", name="article_delete", methods={"DELETE"})
+     * @IsGranted("ROLE_AUTHOR")
      */
     public function delete(Request $request, Article $article): Response
     {
+        $this->denyAccessUnlessGranted('DELETE', $article);
+
         if ($this->isCsrfTokenValid('delete'.$article->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($article);
